Add an event listener to reset the captcha in case of woocommerce checkout error
On the WooCommerce checkout page, if a user validates the reCaptcha but misses a mandatory field, WooCommerce will trigger a checkout error. Problem is: The reCaptcha stays checked but will fails if the order is submitted again.
The only solutions are either to refresh the page (losing the information entered by the user) or wait for the reCaptcha validity to time out.
The proposed change registers an event to reset the object in case of a WC checkout_error event.